refactor(api): extract blob streaming into pkg/api/blobstream.go
Move handleBlobStream and streamBlobFromUpstream out of pkg/api/routes.go
into a dedicated file so the storage/streaming plumbing does not live next
to the route handlers. Pure move, no behavior change.
